requirements.txt
art
python-dotenv
requests